Finished controller methods
Have to finish unit tests for service methods

+++ PetShelter-Backend/src/main/java/ca/mcgill/ecse321/petshelter/service/PetShelterService.java
@@ -2,27 +2,34 @@ package ca.mcgill.ecse321.petshelter.service;
 
 import java.sql.Date;
 import java.util.ArrayList;
-
-import java.util.Collection;
-import java.util.HashSet;
 import java.util.Iterator;
-
 import java.util.List;
-import java.util.Optional;
-import java.util.Set;
-
-// Used for validation
-import java.util.regex.Matcher;
 import java.util.regex.Pattern;
-import java.util.concurrent.TimeUnit;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import ca.mcgill.ecse321.petshelter.model.*;
-import ca.mcgill.ecse321.petshelter.dao.*;
 import ca.mcgill.ecse321.petshelter.ErrorMessages;
+import ca.mcgill.ecse321.petshelter.dao.AdminRepository;
+import ca.mcgill.ecse321.petshelter.dao.ApplicationRepository;
+import ca.mcgill.ecse321.petshelter.dao.ClientRepository;
+import ca.mcgill.ecse321.petshelter.dao.CommentRepository;
+import ca.mcgill.ecse321.petshelter.dao.DonationRepository;
+import ca.mcgill.ecse321.petshelter.dao.MessageRepository;
+import ca.mcgill.ecse321.petshelter.dao.PostingRepository;
+import ca.mcgill.ecse321.petshelter.dao.ProfileRepository;
+import ca.mcgill.ecse321.petshelter.model.Admin;
+import ca.mcgill.ecse321.petshelter.model.Application;
+import ca.mcgill.ecse321.petshelter.model.ApplicationStatus;
+import ca.mcgill.ecse321.petshelter.model.Client;
+import ca.mcgill.ecse321.petshelter.model.Comment;
+import ca.mcgill.ecse321.petshelter.model.Donation;
+import ca.mcgill.ecse321.petshelter.model.HomeType;
+import ca.mcgill.ecse321.petshelter.model.IncomeRange;
+import ca.mcgill.ecse321.petshelter.model.Message;
+import ca.mcgill.ecse321.petshelter.model.Posting;
+import ca.mcgill.ecse321.petshelter.model.Profile;
 
 @Service
 public class PetShelterService {
@@ -45,8 +52,8 @@ public class PetShelterService {
 	private ProfileRepository profileRepository;
 
 	@Transactional
-	public Client createClient(Date dob, String email, String password, String phoneNumber, 
-							   String address, String firstName, String lastName) {
+	public Client createClient(Date dob, String email, String password, String phoneNumber, String address,
+			String firstName, String lastName) {
 
 		// Checking if client exists already
 		try {
@@ -264,8 +271,10 @@ public class PetShelterService {
 	}
 
 	/**
-	 * Method to be used when updating a client profile. From the frontend, the current information will be entered
-	 * and the user can modify it and confirm it.
+	 * Method to be used when updating a client profile. From the frontend, the
+	 * current information will be entered and the user can modify it and confirm
+	 * it.
+	 * 
 	 * @param client
 	 * @param password
 	 * @param phoneNumber
@@ -278,7 +287,7 @@ public class PetShelterService {
 	@Transactional
 	public Client updateClientProfile(Client client, String password, String phoneNumber, String address,
 			String firstName, String lastName, Date dob) {
-		
+
 		client.setAddress(address);
 		client.setDateOfBirth(dob);
 		client.setPassword(password);
@@ -288,30 +297,31 @@ public class PetShelterService {
 		return client;
 
 	}
-	
+
 	/**
 	 * method to get a donation with the id, returns null if doesnt exist.
+	 * 
 	 * @param Id
 	 * @return donation if exists or null
 	 */
 	@Transactional
 	public Donation getDonationbyId(Integer Id) {
-		
-		Donation donation=new Donation();
+
+		Donation donation = new Donation();
 		try {
 			donation = donationRepository.findById(Id).get();
 			return donation;
-		}
-		catch(Exception e) {
+		} catch (Exception e) {
 			return null;
 		}
-		
+
 	}
 
 	/**
 	 * Method to use when a donation is being sent, the amount needs to be an
-	 * integer.
-	 * Method to use when a donation is being sent, the amount needs to be an integer.
+	 * integer. Method to use when a donation is being sent, the amount needs to be
+	 * an integer.
+	 * 
 	 * @param amount
 	 * @param client
 	 * @param date
@@ -320,24 +330,21 @@ public class PetShelterService {
 	@Transactional
 	public Donation sendDonation(Integer amount, Client client, Date date) {
 
+		if (amount <= 0) {
+			throw new IllegalArgumentException(ErrorMessages.negAmount);
+		}
 
-		
-			if(amount<=0) {
-				throw new IllegalArgumentException(ErrorMessages.negAmount);
-			}
-		
-		if(client == null) {
+		if (client == null) {
 			throw new IllegalArgumentException(ErrorMessages.accountDoesNotExist);
 		}
-		if(!client.getIsLoggedIn()) {
-			
+		if (!client.getIsLoggedIn()) {
+
 			throw new IllegalArgumentException(ErrorMessages.notLoggedIn);
-			
+
 		}
-		if(date.before(client.dateOfBirth)) {
+		if (date.before(client.dateOfBirth)) {
 			throw new IllegalArgumentException(ErrorMessages.DateBefDOB);
 		}
-
 
 		Donation donation = new Donation();
 		donation.setAmount(amount);
@@ -348,9 +355,10 @@ public class PetShelterService {
 	}
 
 	/**
-	 * Method used when a message is being sent to admin, the content is checked,
-	 * if the content is repeated too frequently, the message will not be sent.
-	 * To avoid spamming the admin.
+	 * Method used when a message is being sent to admin, the content is checked, if
+	 * the content is repeated too frequently, the message will not be sent. To
+	 * avoid spamming the admin.
+	 * 
 	 * @param admin
 	 * @param client
 	 * @param content
@@ -403,39 +411,40 @@ public class PetShelterService {
 		messageRepository.save(message);
 		return message;
 	}
-	
+
 	/**
-	 * This method returns all the messages a client sent since it created its account.
+	 * This method returns all the messages a client sent since it created its
+	 * account.
+	 * 
 	 * @param client
 	 * @return List<Message>, the list of all messages of the client
 	 */
 	@Transactional
-	public List<Message> getClientMessages(Client client){
-		
-		if(client == null) {
+	public List<Message> getClientMessages(Client client) {
+
+		if (client == null) {
 			throw new IllegalArgumentException(ErrorMessages.accountDoesNotExist);
 		}
-		if(client.getMessages().size() == 0) {
+		if (client.getMessages().size() == 0) {
 			throw new IllegalArgumentException(ErrorMessages.ClientHasNoMessages);
 		}
 		return toList(client.getMessages());
-		
+
 	}
-	
-	
+
 	/**
 	 * Service method to get all the donations of a client.
+	 * 
 	 * @param client
 	 * @return list of donation with all the donations of a client
 	 */
 	@Transactional
-	public List<Donation> getClientDonations(Client client){
-		if(client == null) {
+	public List<Donation> getClientDonations(Client client) {
+		if (client == null) {
 			throw new IllegalArgumentException(ErrorMessages.accountDoesNotExist);
 		}
 		return toList(client.getDonations());
 	}
-	
 
 	@Transactional
 	public Comment commentOnPosting(Profile profile, Posting posting, String content, Date date) {
@@ -447,10 +456,10 @@ public class PetShelterService {
 		if (profile == null) {
 			throw new IllegalArgumentException(ErrorMessages.invalidProfile);
 		}
-		if(!profile.getIsLoggedIn()) {
+		if (!profile.getIsLoggedIn()) {
 			throw new IllegalArgumentException(ErrorMessages.invalidProfileNotLoggedIn);
 		}
-		//check content is not just white spaces
+		// check content is not just white spaces
 		String contentWhiteSpaceCheck = content.trim();
 		if (content == null || contentWhiteSpaceCheck == "" || contentWhiteSpaceCheck == null) {
 			throw new IllegalArgumentException(ErrorMessages.invalidContentComment);
@@ -458,10 +467,10 @@ public class PetShelterService {
 		if (date == null) {
 			throw new IllegalArgumentException(ErrorMessages.invalidDateComment);
 		}
-		if(date.before(profile.getDateOfBirth())) {
+		if (date.before(profile.getDateOfBirth())) {
 			throw new IllegalArgumentException(ErrorMessages.invalidDateCommentProfile);
 		}
-		if(date.before(posting.getDate())) {
+		if (date.before(posting.getDate())) {
 			throw new IllegalArgumentException(ErrorMessages.invalidDateCommentPosting);
 		}
 
@@ -476,38 +485,37 @@ public class PetShelterService {
 		comment = commentRepository.save(comment);
 		return comment;
 	}
-	
+
 	@Transactional
-	public List<Comment> getComments(Posting posting){
-		if(posting == null) {
+	public List<Comment> getComments(Posting posting) {
+		if (posting == null) {
 			throw new IllegalArgumentException(ErrorMessages.invalidPosting);
 		}
 		List<Comment> allComments = toList(commentRepository.findAll());
 		List<Comment> comments = new ArrayList<Comment>();
-		for(Comment comment: allComments) {
-			if(comment.getPosting().equals(posting)) {
+		for (Comment comment : allComments) {
+			if (comment.getPosting().equals(posting)) {
 				String contentWhiteSpaceCheck = comment.getContent().trim();
-				if(comment.getContent() == null || contentWhiteSpaceCheck == "" || contentWhiteSpaceCheck == null) {
+				if (comment.getContent() == null || contentWhiteSpaceCheck == "" || contentWhiteSpaceCheck == null) {
 					throw new IllegalArgumentException(ErrorMessages.invalidContentComment);
 				}
-				if(comment.getDate() == null) {
+				if (comment.getDate() == null) {
 					throw new IllegalArgumentException(ErrorMessages.invalidDateComment);
 				}
-				if(comment.getDate().before(comment.getProfile().getDateOfBirth())) {
+				if (comment.getDate().before(comment.getProfile().getDateOfBirth())) {
 					throw new IllegalArgumentException(ErrorMessages.invalidDateCommentProfile);
 				}
-				if(comment.getDate().before(comment.getPosting().getDate())) {
+				if (comment.getDate().before(comment.getPosting().getDate())) {
 					throw new IllegalArgumentException(ErrorMessages.invalidDateCommentPosting);
 				}
-				
-				//add only valid comments that are on that posting
-				
+
+				// add only valid comments that are on that posting
+
 				comments.add(comment);
 			}
 		}
 		return comments;
 	}
-	
 
 	@Transactional
 	public List<Posting> getOpenPostings() {
@@ -561,7 +569,7 @@ public class PetShelterService {
 		if (!profile.getIsLoggedIn()) {
 			throw new IllegalArgumentException(ErrorMessages.invalidLoggedIn);
 		}
-		
+
 		Posting posting = new Posting();
 		posting.setProfile(profile);
 		posting.setComment(null);
@@ -572,19 +580,23 @@ public class PetShelterService {
 		posting.setPetBreed(breed);
 		posting.setPicture(picture);
 		posting.setDescription(reason);
-		posting.setId(profile.getEmail().hashCode()*postDate.hashCode());
+		posting.setId(profile.getEmail().hashCode() * postDate.hashCode());
 		posting = postingRepository.save(posting);
 		return posting;
 	}
 
 	@Transactional
-	public Posting deletePosting(Posting posting) { 
+	public Posting deletePosting(Posting posting) {
 		if (posting == null) {
 			throw new IllegalArgumentException(ErrorMessages.invalidPosting);
 		}
+		if (!posting.getProfile().getIsLoggedIn()) {
+			throw new IllegalArgumentException(ErrorMessages.invalidLoggedIn);
+		}
+		
 		postingRepository.delete(posting);
 		return posting;
-		
+
 	}
 
 	@Transactional
@@ -609,7 +621,10 @@ public class PetShelterService {
 		if (reason == null || reason.equals("") || reason.length() >= 1000) {
 			throw new IllegalArgumentException(ErrorMessages.invalidReason);
 		}
-		
+		if (!posting.getProfile().getIsLoggedIn()) {
+			throw new IllegalArgumentException(ErrorMessages.invalidLoggedIn);
+		}
+
 		posting.setPetName(petName);
 		posting.setPetDateOfBirth(dob);
 		posting.setPetBreed(breed);
@@ -618,18 +633,18 @@ public class PetShelterService {
 		posting = postingRepository.save(posting);
 		return posting;
 	}
-	
+
 	@Transactional
 	public Posting getPosting(String email, Date date) {
-		if(email == null) {
+		if (email == null) {
 			throw new IllegalArgumentException(ErrorMessages.invalidEmail);
 		}
-		if(date == null) {
+		if (date == null) {
 			throw new IllegalArgumentException(ErrorMessages.invalidDate);
 		}
-		
-		Posting posting = postingRepository.findPostingById(email.hashCode()*date.hashCode());
-		if(posting == null) {
+
+		Posting posting = postingRepository.findPostingById(email.hashCode() * date.hashCode());
+		if (posting == null) {
 			throw new IllegalArgumentException(ErrorMessages.postingDoesNotExist);
 		}
 		return posting;
@@ -637,20 +652,21 @@ public class PetShelterService {
 
 	@Transactional
 	public Application getApplication(String applicant_email, Posting posting) {
-		if(applicant_email == null) {
+		if (applicant_email == null) {
 			throw new IllegalArgumentException(ErrorMessages.invalidEmail);
 		}
-		if(posting == null) {
+		if (posting == null) {
 			throw new IllegalArgumentException(ErrorMessages.invalidPosting);
 		}
-		
-		Application application = applicationRepository.findApplicationById(applicant_email.hashCode() * posting.hashCode()); 
-		if(application == null) {
+
+		Application application = applicationRepository
+				.findApplicationById(applicant_email.hashCode() * posting.hashCode());
+		if (application == null) {
 			throw new IllegalArgumentException(ErrorMessages.applicationDoesNotExist);
 		}
 		return application;
 	}
-	
+
 	public List<Application> getPostingApplications(Posting posting) {
 		if (posting == null) {
 			throw new IllegalArgumentException(ErrorMessages.invalidPosting);
@@ -689,8 +705,8 @@ public class PetShelterService {
 		applicationRepository.save(application);
 		application = applicationRepository.save(application);
 
-		for(Application a : getPostingApplications(application.getPosting())) {
-			if(!(a.equals(application))) {
+		for (Application a : getPostingApplications(application.getPosting())) {
+			if (!(a.equals(application))) {
 				a.setStatus(ApplicationStatus.rejected);
 				applicationRepository.save(a);
 			}
@@ -708,7 +724,7 @@ public class PetShelterService {
 		if (posting == null) {
 			throw new IllegalArgumentException(ErrorMessages.invalidPosting);
 		}
-		if(client.getEmail().equals(posting.getProfile().getEmail())){
+		if (client.getEmail().equals(posting.getProfile().getEmail())) {
 			throw new IllegalArgumentException(ErrorMessages.selfApplication);
 		}
 		if (homeType == null) {
